+++ src/phaser/scenes/create/Level/BaseObjectGroupLayerManager.ts
@@ -1,0 +1,147 @@
+import Phaser from "phaser"
+
+import BaseToolboxManager from "./BaseToolboxManager"
+import type { default as Level } from "."
+
+/** Gap, in pixels, between the rotate and delete buttons. */
+const BUTTON_GAP = 16
+
+/** Margin, in pixels, between the placed object and the button stack. */
+const BUTTON_STACK_MARGIN = 8
+
+export type Placed<ID extends number, VariantKey extends string> = {
+  id: ID
+  variantKey: VariantKey
+  obj: Phaser.GameObjects.Image
+}
+
+/**
+ * Shared logic for tools that place objects and let the player select one to
+ * rotate or delete it via a floating button stack. `Key` identifies a placed
+ * object (e.g. a tile for road objects, or the object itself for free ones).
+ */
+export default abstract class BaseObjectGroupLayerManager<
+  ID extends number,
+  VariantKey extends string,
+  Key,
+> extends BaseToolboxManager<ID> {
+  /** The key of the currently selected placed object. */
+  protected selected: Key | null = null
+
+  private readonly button: {
+    /** The row of buttons shown below the selected object. */
+    stack: Phaser.GameObjects.Stack
+    /** The delete button shown in the button stack. */
+    delete: Phaser.GameObjects.FloatingActionButton
+    /** The rotate button shown in the button stack. */
+    rotate: Phaser.GameObjects.FloatingActionButton
+  }
+
+  constructor(level: Level) {
+    super(level)
+
+    const deleteButton = this.createDeleteButton(level)
+    const rotateButton = this.createRotateButton(level)
+    this.button = {
+      rotate: rotateButton,
+      delete: deleteButton,
+      stack: level.add
+        .stack(0, 0, [rotateButton, deleteButton], {
+          direction: "row",
+          gap: BUTTON_GAP,
+        })
+        .setVisible(false),
+    }
+
+    // The tooltips must be created after the buttons are added to the stack
+    // so that the tooltips can observe the stack's visibility.
+    level.add.tooltip("Rotate", rotateButton)
+    level.add.tooltip("Delete", deleteButton)
+  }
+
+  /** Checks if this manager has a placed object identified by `key`. */
+  isOccupied(key: Key): boolean {
+    return this.getPlaced(key) !== null
+  }
+
+  /** Returns the object placed at `key`, if any. */
+  protected abstract getPlaced(key: Key): Placed<ID, VariantKey> | null
+
+  /** Returns the valid variant keys for the object identified by `key`. */
+  protected abstract validVariantKeys(key: Key, id: ID): VariantKey[]
+
+  /** Removes the placed object identified by `key`, if any. */
+  protected abstract remove(key: Key): void
+
+  /** Rotates the placed object identified by `key` to its next valid variant. */
+  protected abstract rotate(key: Key): void
+
+  /** Show/hide and position the ghost based on the tile under the pointer. */
+  protected abstract handleGhost(pointer?: Phaser.Input.Pointer): void
+
+  /** Visually highlights the object identified by `key` as selected. */
+  protected abstract highlightSelection(key: Key): void
+
+  /** Clears the selection highlight left by `key`. */
+  protected abstract clearSelectionHighlight(key: Key): void
+
+  /** Checks if `a` and `b` identify the same placed object. */
+  protected sameKey(a: Key, b: Key) {
+    return (a as unknown) === (b as unknown)
+  }
+
+  private createDeleteButton({ add }: Level) {
+    const onPointerUp: Phaser.Input.Events.GameObjectPointerUp = pointer => {
+      if (this.selected !== null) this.remove(this.selected)
+      this.handleGhost(pointer)
+    }
+
+    return add
+      .fab(0, 0, "delete-icon", 0xff0000, 0xc0392b)
+      .on(Phaser.Input.Events.POINTER_UP, onPointerUp)
+  }
+
+  private createRotateButton({ add }: Level) {
+    const onPointerUp: Phaser.Input.Events.GameObjectPointerUp = () => {
+      if (this.selected !== null) this.rotate(this.selected)
+    }
+
+    return add
+      .fab(0, 0, "rotate-right-icon", 0x2196f3, 0x1565c0)
+      .on(Phaser.Input.Events.POINTER_UP, onPointerUp)
+  }
+
+  protected select(key: Key) {
+    if (this.selected !== null && this.sameKey(this.selected, key)) return
+    this.deselect()
+    this.selected = key
+    this.highlightSelection(key)
+
+    const placed = this.getPlaced(key)
+    if (!placed) return
+
+    // Use the axis-aligned world bounds so the buttons don't rotate with
+    // the placed object.
+    const bounds = placed.obj.getBounds()
+
+    const validKeys = this.validVariantKeys(key, placed.id)
+    this.button.rotate.setVisible(validKeys.length > 1)
+
+    this.button.stack
+      .setPosition(
+        bounds.centerX - this.button.stack.displayWidth / 2,
+        bounds.bottom +
+          BUTTON_STACK_MARGIN +
+          this.button.stack.displayHeight / 2,
+      )
+      .setVisible(true)
+  }
+
+  protected deselect() {
+    if (this.selected === null) return
+    const key = this.selected
+    this.selected = null
+    this.clearSelectionHighlight(key)
+    this.button.stack.setVisible(false)
+  }
+}
