+++ src/phaser/scenes/create/Level/SceneryManager.ts
@@ -1,227 +1,34 @@
-import Phaser from "phaser"
-
-import * as layers from "../../../layers"
 import * as objects from "../../../layers/objectGroup/objects"
 import type * as tilesets from "../../../tilesets"
-import { Events, Variables } from "../../../globals"
-import BaseToolboxManager from "./BaseToolboxManager"
+import BaseFreeObjectManager from "./BaseFreeObjectManager"
 import type { default as Level } from "."
 
-type Drag = {
-  obj: Phaser.GameObjects.Image
-  start: { x: number; y: number }
-  listeners: { on: () => void; off: () => void }
-}
-
-export default class extends BaseToolboxManager {
-  /** The maximum number of scenery objects that can be added to the level. */
-  private readonly maxLength = 50
-
-  /** The currently selected scenery object. */
-  private selectedObject: Phaser.GameObjects.Image | null = null
-
-  /** The currently dragged scenery object. */
-  private drag: Drag | null = null
-
-  /** The semi-transparent preview image. */
-  private ghost: {
-    obj: Phaser.GameObjects.Image
-    id: tilesets.scenery.ID
-  } | null = null
-
-  /** The delete button shown next to the active object. */
-  private deleteButton: Phaser.GameObjects.FloatingActionButton
-
-  /** Maps each placed scenery object to the tileset id it was created from. */
-  private readonly objectIds = new WeakMap<
-    Phaser.GameObjects.Image,
-    tilesets.scenery.ID
-  >()
-
+export default class extends BaseFreeObjectManager<
+  objects.scenery.Name,
+  tilesets.scenery.ID,
+  ""
+> {
   constructor(level: Level) {
     super(level)
 
-    level.setVariable("sceneryObjectCount", this.scenery.length)
+    level.setVariable("sceneryObjectCount", this.placedObjects.length)
     level.setVariable("maxSceneryObjectCount", this.maxLength)
-
-    this.deleteButton = this.createDeleteButton(level)
-    this.registerEventListeners(level)
   }
 
-  private createDeleteButton({ add }: Level) {
-    const onPointerUp: Phaser.Input.Events.GameObjectPointerUp = pointer => {
-      if (this.selectedObject) this.delete(this.selectedObject)
-      this.handleGhost(pointer)
-    }
-
-    return add
-      .fab(0, 0, "delete-icon", 0xff0000, 0xc0392b)
-      .on(Phaser.Input.Events.POINTER_UP, onPointerUp)
-      .setVisible(false)
+  protected get box() {
+    return "scenery" as const
   }
 
-  private registerEventListeners({ game, input, events }: Level) {
-    const onAddRoad: Phaser.Events.AddRoad = (...args) =>
-      this.onAddRoad(...args)
-    game.events.on(Events.ADD_ROAD, onAddRoad)
-
-    const onAddEndpoint: Phaser.Events.AddEndpoint = (...args) =>
-      this.onAddEndpoint(...args)
-    game.events.on(Events.ADD_ENDPOINT, onAddEndpoint)
-
-    const onReactSetVariable: Phaser.Events.ReactSetVariable = (...args) =>
-      this.onReactSetVariable(...args)
-    game.events.on(Events.REACT_SET_VARIABLE, onReactSetVariable)
-
-    // Phaser fires the scene-level POINTER_DOWN with currentlyOver BEFORE the
-    // individual game-object POINTER_DOWN events, so we can inspect what is
-    // under the cursor here without needing a separate flag.
-    const onPointerDown: Phaser.Input.Events.PointerDown<
-      Phaser.GameObjects.Image
-    > = (...args) => this.onPointerDown(...args)
-    input.on(Phaser.Input.Events.POINTER_DOWN, onPointerDown)
-
-    const onPointerMove: Phaser.Input.Events.PointerMove<
-      Phaser.GameObjects.Image
-    > = (...args) => this.onPointerMove(...args)
-    input.on(Phaser.Input.Events.POINTER_MOVE, onPointerMove)
-
-    const onGameOut: Phaser.Input.Events.GameOut = (...args) =>
-      this.onGameOut(...args)
-    input.on(Phaser.Input.Events.GAME_OUT, onGameOut)
-
-    events.on(Phaser.Scenes.Events.SHUTDOWN, () => {
-      game.events.off(Events.ADD_ROAD, onAddRoad)
-      game.events.off(Events.ADD_ENDPOINT, onAddEndpoint)
-      game.events.off(Events.REACT_SET_VARIABLE, onReactSetVariable)
-      input.off(Phaser.Input.Events.POINTER_DOWN, onPointerDown)
-      input.off(Phaser.Input.Events.POINTER_MOVE, onPointerMove)
-      input.off(Phaser.Input.Events.GAME_OUT, onGameOut)
-    })
+  protected get layerName() {
+    return "ObjectGroup.SCENERY" as const
   }
 
-  private get scenery() {
-    return this.level.layers["ObjectGroup.SCENERY"]
+  /** The maximum number of scenery objects that can be added to the level. */
+  protected get maxLength() {
+    return 50
   }
 
-  private get endpoints() {
-    return this.level.layers["ObjectGroup.ENDPOINTS"]
-  }
-
-  private get tool() {
-    return this.level.toolbox?.box === "scenery"
-      ? this.level.toolbox.tool
-      : null
-  }
-
-  /**
-   * Switches the active toolbox to scenery (if it isn't already), so React
-   * mirrors the change and other managers deactivate.
-   */
-  private claimToolbox(id: tilesets.scenery.ID) {
-    if (!this.tool)
-      this.level.setVariable("toolbox", { box: "scenery", tool: id })
-  }
-
-  /** Check if a scenery object is overlapping an endpoint. */
-  private sceneryIsOverlappingEndpoint(
-    obj: Phaser.GameObjects.Image,
-    endpoint: Phaser.GameObjects.Image,
-    bounds = obj.getBounds(),
-  ) {
-    return (
-      obj !== endpoint &&
-      Phaser.Geom.Rectangle.Overlaps(bounds, endpoint.getBounds())
-    )
-  }
-
-  /** Check if a scenery object can overlap another scenery object. */
-  private sceneryCanOverlapScenery(
-    obj1: Phaser.GameObjects.Image,
-    obj2: Phaser.GameObjects.Image,
-    obj1Bounds = obj1.getBounds(),
-  ) {
-    if (obj1 === obj2) return true
-
-    const obj2Bounds = obj2.getBounds()
-
-    // The centre of one object can never be inside the bounds of another.
-    if (obj2Bounds.contains(obj1Bounds.centerX, obj1Bounds.centerY))
-      return false
-
-    // The bounds of the objects don't overlap - no need for further checks.
-    if (!Phaser.Geom.Rectangle.Overlaps(obj1Bounds, obj2Bounds)) return true
-
-    // 2 objects that are both below ground can't overlap.
-    if (
-      obj1.depth === objects.Depths.BELOW_GROUND &&
-      obj2.depth === objects.Depths.BELOW_GROUND
-    )
-      return false
-
-    // A solid object can't overlap another object unless one is above ground
-    // and the other is below ground.
-    if (
-      (objects.getDensity(obj1.name) === objects.Densities.SOLID ||
-        objects.getDensity(obj2.name) === objects.Densities.SOLID) &&
-      !(
-        (obj1.depth === objects.Depths.BELOW_GROUND &&
-          obj2.depth === objects.Depths.ABOVE_GROUND) ||
-        (obj1.depth === objects.Depths.ABOVE_GROUND &&
-          obj2.depth === objects.Depths.BELOW_GROUND)
-      )
-    )
-      return false
-
-    return true // The objects are allowed to overlap.
-  }
-
-  private sceneryIsOverlappingObject(obj: Phaser.GameObjects.Image): boolean
-  private sceneryIsOverlappingObject(
-    obj: Phaser.GameObjects.Image,
-    x: number,
-    y: number,
-  ): boolean
-  /** Check if a scenery object is overlapping any other object. */
-  private sceneryIsOverlappingObject(
-    obj: Phaser.GameObjects.Image,
-    x = obj.x,
-    y = obj.y,
-  ) {
-    const bounds = obj.getRelativeBounds(x, y)
-
-    return (
-      // Check if the object overlaps any other scenery object.
-      this.scenery.some(s => !this.sceneryCanOverlapScenery(obj, s, bounds)) ||
-      // Check if the object overlaps any endpoint object.
-      this.endpoints.some(e =>
-        this.sceneryIsOverlappingEndpoint(obj, e, bounds),
-      )
-    )
-  }
-
-  /** Check if a scenery object is overlapping a road tile. */
-  private sceneryIsOverlappingRoad(
-    obj: Phaser.GameObjects.Image,
-    x: number,
-    y: number,
-  ) {
-    const bounds = obj.getRelativeBounds(x, y)
-
-    const tiles = this.level.tilemap.getTilesWithinWorldXY(
-      bounds.x,
-      bounds.y,
-      bounds.width,
-      bounds.height,
-      { isNotEmpty: true },
-      undefined,
-      layers.Names.Tile.ROAD,
-    )
-
-    return !!tiles && tiles.length > 0
-  }
-
-  private getFactory(id: tilesets.scenery.ID) {
+  protected getFactory(id: tilesets.scenery.ID) {
     return objects.getFactory<
       objects.scenery.Name,
       tilesets.scenery.ID,
@@ -229,293 +36,7 @@ export default class extends BaseToolboxManager {
     >(id)
   }
 
-  private add(
-    worldX: number,
-    worldY: number,
-    id: tilesets.scenery.ID,
-  ): Phaser.GameObjects.Image | null {
-    if (this.scenery.length >= this.maxLength) return null
-
-    const factory = this.getFactory(id)
-    if (!factory) return null
-
-    let obj = this.level
-      .addObject("ObjectGroup.SCENERY", factory({ x: 0, y: 0 }))
-      .setInteractive()
-      .setOrigin(0.5, 0.5)
-      .setPosition(worldX, worldY)
-
-    this.objectIds.set(obj, id)
-    this.level.setVariable("sceneryObjectCount", this.scenery.length)
-
-    const onDragStart: Phaser.Input.Events.GameObjectDragStart = () => {
-      this.claimToolbox(id)
-      this.startDrag(obj, {
-        on: () => {
-          obj
-            .on(Phaser.Input.Events.DRAG, onDrag)
-            .on(Phaser.Input.Events.DRAG_END, onDragEnd)
-        },
-        off: () => {
-          obj
-            .off(Phaser.Input.Events.DRAG, onDrag)
-            .off(Phaser.Input.Events.DRAG_END, onDragEnd)
-        },
-      })
-    }
-
-    const onDrag: Phaser.Input.Events.GameObjectDrag = (_, dragX, dragY) =>
-      this.dragTo(obj, dragX, dragY)
-
-    const onDragEnd: Phaser.Input.Events.GameObjectDragEnd = () =>
-      this.endDrag(obj)
-
-    const onPointerOver: Phaser.Input.Events.GameObjectPointerOver = () => {
-      if (this.level.isDragging) return
-      this.level.input.setDefaultCursor("grab")
-    }
-
-    const onPointerOut: Phaser.Input.Events.GameObjectPointerOut = () => {
-      if (this.level.isDragging) return
-      this.level.input.setDefaultCursor("default")
-    }
-
-    const onPointerUp: Phaser.Input.Events.GameObjectPointerUp = () => {
-      this.claimToolbox(id)
-      this.select(obj)
-    }
-
-    obj = obj
-      .on(Phaser.Input.Events.DRAG_START, onDragStart)
-      .on(Phaser.Input.Events.POINTER_OVER, onPointerOver)
-      .on(Phaser.Input.Events.POINTER_OUT, onPointerOut)
-      .on(Phaser.Input.Events.POINTER_UP, onPointerUp)
-
-    this.level.input.setDraggable(obj)
-
-    return obj
-  }
-
-  private delete(obj: Phaser.GameObjects.Image) {
-    if (this.selectedObject === obj) this.deselect()
-    this.level.destroyObject("ObjectGroup.SCENERY", obj)
-    this.level.setVariable("sceneryObjectCount", this.scenery.length)
-  }
-
-  /** Begin dragging the given scenery object from its current position. */
-  private startDrag(
-    obj: Phaser.GameObjects.Image,
-    listeners: Drag["listeners"],
-  ) {
-    this.drag = { obj, start: { x: obj.x, y: obj.y }, listeners }
-    this.setIsDragging(true)
-    this.level.input.setDefaultCursor("grabbing")
-    obj.setScale(1.1)
-    this.deselect()
-    listeners.on()
-  }
-
-  /**
-   * Move the dragged object to the given position, snapping back to where the
-   * drag started if the position is invalid.
-   */
-  private dragTo(obj: Phaser.GameObjects.Image, dragX: number, dragY: number) {
-    if (!this.drag) return
-
-    const [x, y, cursor] =
-      this.sceneryIsOverlappingObject(obj, dragX, dragY) ||
-      this.sceneryIsOverlappingRoad(obj, dragX, dragY)
-        ? [this.drag.start.x, this.drag.start.y, "not-allowed"]
-        : [dragX, dragY, "grabbing"]
-
-    obj.setPosition(x, y)
-    this.level.input.setDefaultCursor(cursor)
-  }
-
-  /** Stop dragging the given scenery object. */
-  private endDrag(obj: Phaser.GameObjects.Image) {
-    if (!this.drag) return
-
-    this.drag.listeners.off()
-    this.drag = null
-    this.setIsDragging(false)
-    this.level.input.setDefaultCursor("grab")
-    obj.setScale(1)
-  }
-
-  /**
-   * Phaser only starts its own drag tracking for objects that were already
-   * interactive when the pointer went down, so a just-placed object won't be
-   * picked up automatically. If the pointer is still held after placing an
-   * object, drive the same drag behaviour manually until it's released.
-   */
-  private dragNewObject(obj: Phaser.GameObjects.Image) {
-    const onPointerMove: Phaser.Input.Events.PointerMove<
-      Phaser.GameObjects.Image
-    > = pointer => this.dragTo(obj, pointer.worldX, pointer.worldY)
-
-    const onPointerUp: Phaser.Input.Events.PointerUp<
-      Phaser.GameObjects.Image
-    > = () => this.endDrag(obj)
-
-    this.startDrag(obj, {
-      on: () => {
-        this.level.input.on(Phaser.Input.Events.POINTER_MOVE, onPointerMove)
-        this.level.input.on(Phaser.Input.Events.POINTER_UP, onPointerUp)
-      },
-      off: () => {
-        this.level.input.off(Phaser.Input.Events.POINTER_MOVE, onPointerMove)
-        this.level.input.off(Phaser.Input.Events.POINTER_UP, onPointerUp)
-      },
-    })
-  }
-
-  private select(obj: Phaser.GameObjects.Image) {
-    if (this.selectedObject === obj) return
-    this.deselect()
-    this.selectedObject = obj
-    obj.setTint(0xaaddff)
-
-    const { x, y } = obj.getTopRight()
-    this.deleteButton
-      .setPosition(x + this.deleteButton.radius, y - this.deleteButton.radius)
-      .setVisible(true)
-  }
-
-  private deselect() {
-    if (!this.selectedObject) return
-    this.selectedObject.clearTint()
-    this.selectedObject = null
-
-    this.deleteButton.setVisible(false)
-  }
-
-  private createGhost(id: tilesets.scenery.ID) {
-    if (this.ghost?.id === id) return
-    this.destroyGhost()
-
-    this.ghost = {
-      id,
-      obj: this.level.add
-        .imageFromTileset(0, 0, id)
-        .setOrigin(0.5, 0.5)
-        .setAlpha(0.5)
-        .setVisible(false),
-    }
-  }
-
-  private handleGhost(
-    pointer?: Phaser.Input.Pointer,
-    currentlyOver: Phaser.GameObjects.Image[] = pointer
-      ? (this.level.input.hitTestPointer(pointer) as Phaser.GameObjects.Image[])
-      : [],
-  ) {
-    if (!this.ghost) return
-
-    // Directly over an existing object or dragging an object.
-    if (!pointer || currentlyOver.length > 0 || this.drag) {
-      this.ghost.obj.setVisible(false)
-      return
-    }
-
-    // Defer to whichever manager owns this tile (it independently shows its
-    // own "grab" cursor) rather than clobbering it with "not-allowed".
-    const nearestTile = this.level.worldToNearestTile(
-      pointer.worldX,
-      pointer.worldY,
-    )
-    if (nearestTile && this.level.isTileOccupied(nearestTile)) {
-      this.ghost.obj.setVisible(false)
-      return
-    }
-
-    // Indirectly over an existing object or road tile.
-    if (
-      this.scenery.length >= this.maxLength ||
-      this.sceneryIsOverlappingObject(
-        this.ghost.obj,
-        pointer.worldX,
-        pointer.worldY,
-      ) ||
-      this.sceneryIsOverlappingRoad(
-        this.ghost.obj,
-        pointer.worldX,
-        pointer.worldY,
-      )
-    ) {
-      this.ghost.obj.setVisible(false)
-      this.level.input.setDefaultCursor("not-allowed")
-      return
-    }
-
-    this.ghost.obj.setPosition(pointer.worldX, pointer.worldY).setVisible(true)
-    this.level.input.setDefaultCursor("grabbing")
-  }
-
-  private destroyGhost() {
-    this.ghost?.obj.destroy()
-    this.ghost = null
-  }
-
-  private onPointerDown: Phaser.Input.Events.PointerDown<Phaser.GameObjects.Image> =
-    (pointer, currentlyOver) => {
-      const tool = this.tool
-      if (!tool) return
-
-      // Don't interfere while another manager is mid-drag.
-      if (this.level.isDragging) return
-
-      // Clicking on any existing interactive object (scenery, delete button, …):
-      // let the individual object's events handle it.
-      if (currentlyOver.length > 0) return
-
-      // Only place if the ghost is visible, meaning the position is valid.
-      if (!this.ghost?.obj.visible) return
-
-      this.deselect()
-      const obj = this.add(pointer.worldX, pointer.worldY, tool)
-
-      // Seamlessly continue into a drag if the pointer is still held down.
-      if (obj && pointer.isDown) this.dragNewObject(obj)
-    }
-
-  private onPointerMove: Phaser.Input.Events.PointerMove<Phaser.GameObjects.Image> =
-    (pointer, currentlyOver) => {
-      if (!this.tool || this.level.isDragging) return
-      this.handleGhost(pointer, currentlyOver)
-    }
-
-  /** Handle the pointer leaving the game canvas. */
-  private onGameOut: Phaser.Input.Events.GameOut = () => {
-    if (this.ghost) this.handleGhost()
-    if (this.selectedObject) this.deselect()
-    if (this.drag) this.endDrag(this.drag.obj)
-  }
-
-  /** When a road is added, delete any overlapping scenery objects. */
-  private onAddRoad: Phaser.Events.AddRoad = ({ col, row }) => {
-    const tile = this.level.tileToBounds({ col, row })
-    if (!tile) return
-
-    for (const obj of [...this.scenery]) {
-      if (this.level.objectOverlapsTile(obj, tile)) this.delete(obj)
-    }
-  }
-
-  private onAddEndpoint: Phaser.Events.AddEndpoint = ({ obj: endpoint }) => {
-    for (const obj of [...this.scenery]) {
-      if (this.sceneryIsOverlappingEndpoint(obj, endpoint)) this.delete(obj)
-    }
-  }
-
-  private onReactSetVariable: Phaser.Events.ReactSetVariable = key => {
-    if (key !== Variables.TOOLBOX) return
-
-    const tool = this.tool
-    if (tool) this.createGhost(tool)
-    else {
-      this.deselect()
-      this.destroyGhost()
-    }
+  protected onPlacedCountChange() {
+    this.level.setVariable("sceneryObjectCount", this.placedObjects.length)
   }
 }
